Manually add polygon/area to spatial data

Allows users to add polygons (e.g. a closure area/zone) to spatial data by passing a vector or matrix containing coordinates.

Usage

edit_map_manual(spat, id, project, coord = NULL)

Arguments

spat

Spatial data to edit.

id

Spatial ID column.The name must be unique.

project

Name of project.

coord

Optional, a numeric vector or matrix with two columns containing the coordinates of a polygon. Users can build polygons point-by-point using the longitude and latitude boxes in the app.

Details

There are two ways to add polygons when using this function: by setting the coord argument or by building a polygon point-by-point. To build a polygon manually, enter each set of coordinates into the longitude and latitude boxes and click "Add new point". A table of coordinates will appear next to the lon/lat boxes. The polygon table will auto-populate if coord contains a vector or matrix. The coordinate table can be edited by double-clicking a cell. When finished, enter a unique ID into the "ID for new polygon" box and click "Add new polygon". Users can view their polygon in the map viewer at the top of the app. To add a new polygon, select "Clear table" and repeat. Click "Save" to save the edited version of the spatial table.

The edited spatial table will be saved to the projects data/spat/ folder using the naming convention "projectNameSpatTable_modn" ("modn" refers to modified version of the spatial table, e.g. "mod2"). Use list_tables("project", type = "spat") to show spatial tables by project and table_view("table", "project") to view or assign to global environment. To edit a map interactively, see edit_map_gui.

See also

Examples

if (FALSE) {
new_area <- matrix(c(-174.6, 55.6, -172.7, 55.6, 
                     -173.1, 54.4, -175.3, 54.8), 
                     ncol = 2, byrow = TRUE)
edit_map_manual(pollockNMFSSpatTable, id = "NMFS_AREA", project = "pollock",
                coord = new_area)
}