Explore starting value parameter space

Shotgun method to find better parameter starting values by exploring starting value parameter space.

Usage

explore_startparams(project, space, dev, startsr = NULL)

Arguments

project

String, name of project.

space

List of length 1 or length equal to the number of models to be evaluated. space is the number of starting value permutations to test (the size of the space to explore). The greater the dev argument, the larger the space argument should be.

dev

List of length 1 or length equal to the number of models to be evaluated. dev refers to how far to deviate from the average parameter values when exploring (random normal deviates). The less certain the average parameters are, the greater the dev argument should be.

startsr

Optional. List, average starting value parameters for revenue/location-specific covariates then cost/distance. The best guess at what the starting value parameters should be (e.g. all ones). Specify starting value parameters for each model if values should be differetn than ones. The number of starting value parameters should correspond to the likelihood and data that you want to test.

Value

Returns three data frames.

newstart:	Chosen starting values with smallest likelihood	saveLLstarts:	Likelihood values for each starting value permutation

Details

Function is used to identify better starting parameters when convergence is an issue. For more details on the likelihood functions or data, see make_model_design. Function calls the model design file and should be used after the make_model_design function is called.
If more than one model is defined in the model design file, then starting parameters must be defined for each model.

Examples