Log metadata to project folder Meta data is saved to the metadata `meta_log.json` file in a project's `doc/` directory.

Log metadata to project folder Meta data is saved to the metadata `meta_log.json` file in a project's `doc/` directory.

Usage

meta_log_call(
  project,
  meta,
  dataset = NULL,
  tab_name,
  tab_type,
  overwrite = FALSE,
  raw = TRUE
)

Arguments

project

Project name.

meta

Meta data list to be saved.

dataset

A data object. Used to create the column description section of the metadata log entry. If NULL, the section is left empty. This may be the case if saving the raw metadata. To add column descriptions, use the meta_form app.

tab_name

The table name as it appears in the FishSET Database (e.g. "projectMainDataTable" if the main table).

tab_type

The table type. Options include "main", "spat" (spatial), "port", "grid" (gridded), and "aux" (auxiliary).

overwrite

Logical, whether to save over existing metadata entry.

raw

Logical, whether raw metadata file is being saved.