Import, create, and edit metadata

metadata_gui allows users to import metadata from various file types, create and save new metadata, and edit metadata in a shiny application. Metadata is stored in the user's project folder.

Usage

metadata_gui()

Details

The app has two tabs: "Create" and "Edit". The Create tab allows users to create new metadata for a selected FishSET table. When a table is loaded, the app creates several text boxes that the user can fill. There are four metadata sections: About, Column Description, Contact Info, and Other.

About

Author The author of the data.

Date created The date data was created.

Date modified The last data the data was modified.

Version The current version of the data.

Confidentiality Whether the data contains confidential information.

Column Description

A text box for each column in the data. Include the data type, unit, and values (if categorical)

.
Contact Info

Person The primary contact.

Organization The primary contact's organization.

Address The primary contact's and/or organization's address.

Phone The primary contact's work phone number.

Email The primary contact's work email.

Other

License License for data.

Citation Citation for data.

Other Other relevant information.

Users can also import a metadata file from the Create tab, for example, an XML, CSV, or JSON file. This gets saved as "raw" metadata and is separate from the user-created metadata. To see a comprehensive list of accepted file types, see parse_meta and read_dat. To extract metadata from a data file (i.e. the data and metadata are both in the same file, but the metadata is not contained within the data itself), use the Reader parameters text box to selectively parse the file (see parse_meta for details).

The Edit tab allows users to view, edit, and/or delete metadata saved to FishSET.

Examples

if (FALSE) {
metadata_gui()
}