Policy change metrics

Policy change metrics

Usage

policy_metrics(
  dat,
  project,
  tripID = "row",
  vesselID,
  catchID,
  datevar = NULL,
  price = NULL
)

Arguments

dat

Primary data containing information on hauls or trips. Table in the FishSET database contains the string 'MainDataTable'.

project

Name of project

tripID

Trip identifier. Can be 'row' or the name or names of variables that define trips. If tripID='row' then each row of the primary dataset is considered to be a unique trip

vesselID

Vessel identifier. Variable name in primary dataset that contains unique vessel identifier.

catchID

Name of variable in primary dataset that contains catch data.

datevar

Name of variable containing date data. Used to split data into years.

price

Name of variable containing data on revenue or price data.

Value

Tables containing basic metrics on effects of proposed zone closures.

Details

The policy change metrics reflect the impact of proposed policies in the absence of changes in fisher behavior. Policy scenarios are defined using zone_closure() function. Percent of vessels is calculated from the unique vessel identifiers grouped by year and zone. Trips are identified using the tripID argument, otherwise each row is assumed to be a trip. If price is not defined then percent of revenue loss will be reported as NA.