Runs policy scenarios

Estimate redistributed fishing effort and welfare loss/gain from changes in policy or change in other factors that influence fisher location choice.

Usage

run_policy(
  project,
  mod.name = NULL,
  policy.name = NULL,
  betadraws = 1000,
  marg_util_income = NULL,
  income_cost = NULL,
  zone.dat = NULL,
  group_var = NULL,
  enteredPrice = NULL,
  expected.catch = NULL,
  use.scalers = FALSE,
  scaler.func = NULL
)

Arguments

project

Name of project

mod.name

Model name. Argument can be the name of the model or the name can be pulled the `modelChosen` table. Leave mod.name empty to use the name of the saved `best` model. If more than one model is saved, mod.name should be the numeric indicator of which model to use. Use table_view("modelChosen", project) to view a table of saved models.

policy.name

List of policy scenario names created in zone_closure function

betadraws

Integer indicating the number of times to run the welfare simulation. Default value is betadraws = 1000

marg_util_income

For conditional and zonal logit models. Name of the coefficient to use as marginal utility of income.

income_cost

For conditional and zonal logit models. Logical indicating whether the coefficient for the marginal utility of income relates to cost (TRUE) or revenue (FALSE).

zone.dat

Variable in primary data table that contains unique zone ID.

group_var

Categorical variable from primary data table to group welfare outputs.

enteredPrice

Price data. Leave as NULL if using price data from primary dataset.

expected.catch

Required for conditional logit (logit_c) model. Name of expected catch table to use. Can be the expected catch from the short-term scenario (short), the medium-term scenario (med), the long-term scenario (long), or the user-defined temporal parameters (user).

use.scalers

Input for create_model_input(). Logical, should data be normalized? Defaults to FALSE. Rescaling factors are the mean of the numeric vector unless specified with scaler.func.

scaler.func

Input for create_model_input(). Function to calculate rescaling factors.

Details

run_policy is a wrapper function for model_prediction and welfare_predict. model_prediction estimates redistributed fishing effort after policy changes, and welfare_predict simulates welfare loss/gain.